pyproject.toml
[project]
name = "Resume-builder"
version = "0.1.0"

readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.118.0",
    "pydantic-settings>=2.11.0",
    "python-json-logger>=4.0.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
    "gunicorn>=23.0.0",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.10",
    "jinja2>=3.1.6",
    "slowapi>=0.1.9",
    "httpx>=0.28.1",
    "pydantic[email]>=2.12.0",
    "argon2-cffi>=25.1.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "pypdf2>=3.0.1",
    "openai>=1.10.0",
    "google-adk>=1.20.0",
    "types-pyyaml>=6.0.12.20250915",
]

[dependency-groups]
dev= [
    "codespell==2.4.1",
    "mypy>=1.18.2",
    "ruff>=0.14.0",
    "pytest>=8.4.1",
    "pyright>=1.1.407",
]


[tool.mypy]
plugins = ["pydantic.mypy"]
exclude = ["adk_test_projects", "tests", "reproduce_issue.py"]

[tool.ruff]
exclude = ["adk_test_projects", "tests", "reproduce_issue.py"]

[tool.ruff.lint]
ignore = ["E501"]
extend-select = ["E", "I", "F", "G", "N", "T201"]

